category edits only when admin

+++ category-dashboard.php
@@ -6,26 +6,38 @@
 <?php
 	$category_name = "";
 	$category_err  = "";
+	$user_id 	   = $_SESSION['id'];
+	$sql 		   = "SELECT is_admin FROM users WHERE $user_id = users.id";
+	$result 	   = mysqli_query($conn, $sql);
 
-	if($_SERVER["REQUEST_METHOD"] == "POST"){
-		if(!empty($_POST["new_category"])){
-			$category_name = $_POST["new_category"];
-			$sql_request = "INSERT INTO categories(title) VALUES('" . $category_name . "') ";
-
-			if ($conn->query($sql_request) === FALSE) {
-				echo "Error: " . $sql_request . "<br>" . $conn->error;
-			}
+	if ($result->num_rows > 0) {
+		$row = $result->fetch_assoc();
+		if(empty($row)){
+			echo "0 results";
 		}
 	}
-
-	if($_SERVER["REQUEST_METHOD"] == "GET"){
-		if(!empty($_GET['cat_id'])){
-			$delete_category_id = $_GET['cat_id'];
-			$delete_request = "DELETE FROM `categories` WHERE id= " . $delete_category_id . " ";
-		
-				
-			if ($conn->query($delete_request) === FALSE) {
-				echo "Error: " . $delete_request . "<br>" . $conn->error;
+	$is_admin = $row['is_admin'];
+	if($row['is_admin'] == 1){
+		if($_SERVER["REQUEST_METHOD"] == "POST"){
+			if(!empty($_POST["new_category"])){
+				$category_name = $_POST["new_category"];
+				$sql_request = "INSERT INTO categories(title) VALUES('" . $category_name . "') ";
+	
+				if ($conn->query($sql_request) === FALSE) {
+					echo "Error: " . $sql_request . "<br>" . $conn->error;
+				}
+			}
+		}
+	
+		if($_SERVER["REQUEST_METHOD"] == "GET"){
+			if(!empty($_GET['cat_id'])){
+				$delete_category_id = $_GET['cat_id'];
+				$delete_request = "DELETE FROM `categories` WHERE id= " . $delete_category_id . " ";
+			
+					
+				if ($conn->query($delete_request) === FALSE) {
+					echo "Error: " . $delete_request . "<br>" . $conn->error;
+				}
 			}
 		}
 	}
@@ -55,7 +67,9 @@
 										<div class="form-field-wrapper">
 											<input type="text" name="new_category" placeholder="Enter Category Name..."/>
 										</div>
-										<button class="button" >Add New</button>
+										<?php if($row['is_admin'] == 1){?>
+											<button class="button" >Add New</button>
+										<?php }; ?>
 									</div>	
 								</form>
 							</div>
@@ -86,7 +100,9 @@
 							</div>
 							<div class="job-secondary centered-content">
 								<div class="job-actions">
+									<?php if(isset($is_admin) && $is_admin){?>
 									<a href="<?php echo $_SERVER["PHP_SELF"]?>?cat_id=<?php echo $row['id']; ?>" class="button button-inline">Delete</a>
+									<?php }; ?>
 								</div>
 							</div>
 						</li>
